fix re-render problem
